Refactor Warden strategy for handling OmniAuth
Move the generic OmniAuth code in the auth0 Warden strategy into a
separate module, so it can be reused in other OmniAuth/Warden
strategies.
